+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Fixes
 
 - rpc: add `getSlot` to config ([#1](https://github.com/lamports-dev/alpamayo/pull/1))
+- rpc: change upstream header to `x-bigtable: disabled` ([#5](https://github.com/lamports-dev/alpamayo/pull/5))
 
 ### Features